Start working on persisting workspaces to db
